[CCUBE-1414][WL] update e-signature typings - barrel export - move internal typings out of main typing file
