+++ src/e-signature/index.ts
@@ -1,1 +1,2 @@
 export * from "./e-signature";
+export * from "./types";
