+++ src/e-signature/types.ts
@@ -1,5 +1,3 @@
-import { Ref } from "react";
-
 export interface EsignatureProps {
     className?: string | undefined;
     id?: string | undefined;
@@ -8,14 +6,4 @@ export interface EsignatureProps {
     /** image dataURL */
     value?: string | undefined;
     onChange?: ((value: string) => void) | undefined;
-}
-
-export interface ESignatureCanvasProps {
-    baseImageDataURL?: string | undefined;
-    forwardedRef?: Ref<ESignatureCanvasRef> | undefined;
-}
-
-export interface ESignatureCanvasRef {
-    clear: VoidFunction;
-    export: () => string;
 }
